admin_panel/requirements.txt
Django
gunicorn
pymysql==1.1.1
python-decouple==3.8
cryptography==42.0.5
whitenoise